backend/requirements.txt
alembic==1.17.0
aniso8601==10.0.1
attrs==25.4.0
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudinary==1.44.1
cryptography==46.0.3
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
greenlet==3.2.4
idna==3.11
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==4.0.1
marshmallow-sqlalchemy==1.4.2
pillow==10.4.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.2.1
python-http-client==3.3.7
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
sendgrid==6.12.5
six==1.17.0
SQLAlchemy==2.0.44
stripe==11.4.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
